Initial realtime alerts support
The main goal of this commit is to provide initial support for realtime alerts
and the NYCT alerts extension. It ended up being a large refactor becuase
adding the second extension made me realize the initial extensions approach
did not scale will. This commit adds a polymorphic system for extensions in
which extensions just satisfy an Extension interface.