Remove grouped stations feature
The grouped stations feature supported interpreting transfers.txt entries
as meaning two stations were in the same complex. In this case a "grouped
station" would be created, with the stations as children. This feature was
ported over from Transiter.

I think this feature was a mistake. Any such serious reprocessing of GTFS
data should ideally happen in the "last mile"; i.e., in the UI for a subway
app. Supporting this feature in middleware like Transiter was a lot of work.