Performance optimize the static parser
This commit is a significant optimization of the GTFS parser. Profiling using
the NYC buses data, this commit reduces the time parse from ~5 seconds to
~3 seconds.

The code changes are pretty significant, unfortunately. The main reason is that
we had to change the API of the CSV subpackage. A lot of time was being spent
doing map lookups to find which column a particular header was associated to.
This commit moves this logic to run once per-column per-file rather than
once per-column per-row.

There are a bunch of small other changes too.

Finally, some error handling code was updated because some errors were being
swallowed and not bubbled up.