Add trip journaler
A trip journal is built from a sequence of GTFS realtime
messages and contains a historical record of where the trips
stopped and at what times. Ideally I would eventually clean
up the code and test it and document it more.

This code was moved from the subwaydata.nyc repo.